requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

discord.py==2.4.0
python-dotenv==1.0.1
requests==2.32.3
tabulate==0.9.0
